generic session exit event
