+++ app/src/sessions/index.js
@@ -3,4 +3,3 @@
 export * from './actions'
 export * from './constants'
 export * from './selectors'
-export * from './helpers'
