+++ app/src/sessions/selectors.js
@@ -1,5 +1,6 @@
 // @flow
 import type { State } from '../types'
+import * as Constants from './constants'
 import * as Types from './types'
 
 export const getRobotSessions: (
@@ -27,4 +28,54 @@ export const getRobotSessionOfType: (
       id => sessionsById[id].sessionType === sessionType
     ) ?? null
   return foundSessionId ? sessionsById[foundSessionId] : null
+}
+
+export const getAnalyticsPropsForRobotSessionById: (
+  state: State,
+  robotName: string,
+  sessionId: string
+) => Types.SessionAnalyticsProps | null = (state, robotName, sessionId) => {
+  const session = getRobotSessionById(state, robotName, sessionId)
+  if (!session) return null
+
+  if (session.sessionType === Constants.SESSION_TYPE_CALIBRATION_CHECK) {
+    const { instruments, comparisonsByStep } = session.details
+    const initialModelsByMount: Types.AnalyticsModelsByMount = {}
+    const modelsByMount: Types.AnalyticsModelsByMount = Object.keys(
+      instruments
+    ).reduce(
+      (acc, mount) => ({
+        ...acc,
+        [`${mount.toLowerCase()}PipetteModel`]: instruments[mount].model,
+      }),
+      initialModelsByMount
+    )
+    const initialStepData: Types.CalibrationCheckAnalyticsData = {}
+    const normalizedStepData = Object.keys(comparisonsByStep).reduce(
+      (acc, stepName) => {
+        const {
+          differenceVector,
+          thresholdVector,
+          exceedsThreshold,
+          transformType,
+        } = comparisonsByStep[stepName]
+        return {
+          ...acc,
+          [`${stepName}DifferenceVector`]: differenceVector,
+          [`${stepName}ThresholdVector`]: thresholdVector,
+          [`${stepName}ExceedsThreshold`]: exceedsThreshold,
+          [`${stepName}ErrorSource`]: transformType,
+        }
+      },
+      initialStepData
+    )
+    return {
+      sessionType: session.sessionType,
+      ...modelsByMount,
+      ...normalizedStepData,
+    }
+  } else {
+    // the exited session type doesn't report to analytics
+    return null
+  }
 }
