+++ app/src/sessions/types.js
@@ -220,31 +220,33 @@ export type AnalyticsModelsByMount =
       rightPipetteModel?: string,
     |}
   | {||}
+
+type VectorTuple = [number, number, number]
 export type CalibrationCheckAnalyticsData =
   | {|
-      comparingFirstPipetteHeightDifferenceVector?: string,
-      comparingFirstPipetteHeightThresholdVector?: string,
-      comparingFirstPipetteHeightExceedsThreshold?: string,
+      comparingFirstPipetteHeightDifferenceVector?: VectorTuple,
+      comparingFirstPipetteHeightThresholdVector?: VectorTuple,
+      comparingFirstPipetteHeightExceedsThreshold?: boolean,
       comparingFirstPipetteHeightErrorSource?: string,
-      comparingFirstPipettePointOneDifferenceVector?: string,
-      comparingFirstPipettePointOneThresholdVector?: string,
-      comparingFirstPipettePointOneExceedsThreshold?: string,
+      comparingFirstPipettePointOneDifferenceVector?: VectorTuple,
+      comparingFirstPipettePointOneThresholdVector?: VectorTuple,
+      comparingFirstPipettePointOneExceedsThreshold?: boolean,
       comparingFirstPipettePointOneErrorSource?: string,
-      comparingFirstPipettePointTwoDifferenceVector?: string,
-      comparingFirstPipettePointTwoThresholdVector?: string,
-      comparingFirstPipettePointTwoExceedsThreshold?: string,
+      comparingFirstPipettePointTwoDifferenceVector?: VectorTuple,
+      comparingFirstPipettePointTwoThresholdVector?: VectorTuple,
+      comparingFirstPipettePointTwoExceedsThreshold?: boolean,
       comparingFirstPipettePointTwoErrorSource?: string,
-      comparingFirstPipettePointThreeDifferenceVector?: string,
-      comparingFirstPipettePointThreeThresholdVector?: string,
-      comparingFirstPipettePointThreeExceedsThreshold?: string,
+      comparingFirstPipettePointThreeDifferenceVector?: VectorTuple,
+      comparingFirstPipettePointThreeThresholdVector?: VectorTuple,
+      comparingFirstPipettePointThreeExceedsThreshold?: boolean,
       comparingFirstPipettePointThreeErrorSource?: string,
-      comparingSecondPipetteHeightDifferenceVector?: string,
-      comparingSecondPipetteHeightThresholdVector?: string,
-      comparingSecondPipetteHeightExceedsThreshold?: string,
+      comparingSecondPipetteHeightDifferenceVector?: VectorTuple,
+      comparingSecondPipetteHeightThresholdVector?: VectorTuple,
+      comparingSecondPipetteHeightExceedsThreshold?: boolean,
       comparingSecondPipetteHeightErrorSource?: string,
-      comparingSecondPipettePointOneDifferenceVector?: string,
-      comparingSecondPipettePointOneThresholdVector?: string,
-      comparingSecondPipettePointOneExceedsThreshold?: string,
+      comparingSecondPipettePointOneDifferenceVector?: VectorTuple,
+      comparingSecondPipettePointOneThresholdVector?: VectorTuple,
+      comparingSecondPipettePointOneExceedsThreshold?: boolean,
       comparingSecondPipettePointOneErrorSource?: string,
     |}
   | {||}
